Added .gitignore to ignore the symlinks.
GitHub does not allow to push symlinks that link to pathes in outside of
the repository. In fact, the symlinks to outside the directory are useful to
write documents which span to multiple repositories. ( I need ctrl-X+F
combination on VIM. )

Additionally any files that collide to the existing repositories on the
current organization are also problematic on GitHub pages.

I decided to simply add symlinks to the GitHub pages' repository and
add thoses symlink to .gitignore.